% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NDSWorkshop-package.R
\docType{package}
\name{NDSWorkshop-package}
\alias{NDSWorkshop}
\alias{NDSWorkshop-package}
\title{NDSWorkshop: Nonlinear Data Analysis Functions}
\description{
Several functions that make the analysis of nonlinear time series data are included here. These are for use specifically in the Instats Introduction to Nonlinear Dynamical Systems (NDS) workshop, although they may find use elsewhere.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/barneyricca/NDSWorkshop}
  \item Report bugs at \url{https://github.com/barneyricca/NDSWorkshop/issues}
}

}
\author{
\strong{Maintainer}: Barney Ricca \email{barneyricca@gmail.com} (\href{https://orcid.org/0000-0001-6318-1222}{ORCID})

Authors:
\itemize{
  \item Jonathan Butner
}

}
\keyword{internal}
